% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_common.R
\name{align_annotations}
\alias{align_annotations}
\title{Auto indent mrgsolve code}
\usage{
align_annotations(code)
}
\arguments{
\item{code}{Code lines containing comments to align across}
}
\description{
Auto indent mrgsolve code
}
