% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_common.R
\name{align_tags}
\alias{align_tags}
\title{Auto indent NONMEM code tags}
\usage{
align_tags(code)
}
\arguments{
\item{code}{Code lines containing comments to align across}
}
\description{
Auto indent NONMEM code tags
}
