% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{areTruthy}
\alias{areTruthy}
\title{Multi-argument version of isTruthy}
\usage{
areTruthy(...)
}
\arguments{
\item{...}{Any object}
}
\value{
TRUE if all objects are "truthy", FALSE if at least one is not.
}
\description{
See \code{shiny::\link[shiny:isTruthy]{isTruthy}} for details.
}
