% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bottom_triangle}
\alias{bottom_triangle}
\title{Lower part of a matrix (including diagoncal element)}
\usage{
bottom_triangle(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
A matrix of logicals identifying the elements of x which part of the
lower triangle and the diagonal of x
}
\description{
Lower part of a matrix (including diagoncal element)
}
