% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_covariates.R
\name{check_covariate_table}
\alias{check_covariate_table}
\title{Check covariate definition table}
\usage{
check_covariate_table(table = NULL, check_step = FALSE)
}
\arguments{
\item{table}{a data.frame with expected columns}

\item{check_step}{a logical indicator to check content of Step column or not}
}
\description{
Check covariate definition table
}
