% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_covariates.R
\name{check_incomplete_covariate_table}
\alias{check_incomplete_covariate_table}
\title{Check if covariate definition table contains rows with incomplete data}
\usage{
check_incomplete_covariate_table(table)
}
\arguments{
\item{table}{a data.frame with expected columns}
}
\description{
Check if covariate definition table contains rows with incomplete data
}
