% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_covariates.R
\name{convert_reference_code}
\alias{convert_reference_code}
\title{Convert reference code for univariate model creation}
\usage{
convert_reference_code(code)
}
\arguments{
\item{code}{A character string containing NONMEM code}
}
\description{
Convert reference code for univariate model creation
}
