% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates_server.R
\name{covariates_server}
\alias{covariates_server}
\title{New model module}
\usage{
covariates_server(session, input, output, resources)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}

\item{resources}{A list of internal resources}
}
\description{
New model module
}
