% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_covariates.R
\name{create_univariate_models}
\alias{create_univariate_models}
\title{Generate univariate model code}
\usage{
create_univariate_models(
  code,
  referenceName,
  nThetas,
  table,
  style,
  prefix,
  startNumber,
  path
)
}
\arguments{
\item{code}{A character string containing the code of the reference NONMEM model}

\item{referenceName}{The name of the reference model file}

\item{nThetas}{The number of THETA parameters in the reference NONMEM model}

\item{table}{The table of covariate relationship definition}

\item{style}{Either PsN or standard}

\item{prefix}{If standard style, the prefix to start the name of the univariable
model file with}

\item{startNumber}{If PsN style, the number to start univariate model run files}

\item{path}{The directory in which the univariate models will be saved and run}
}
\description{
Generate univariate model code
}
