% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_correlation_table}
\alias{get_correlation_table}
\title{Convert of covariance matrix into a 0/1 correlation map table}
\usage{
get_correlation_table(x, na_zero = FALSE)
}
\arguments{
\item{x}{A covariance matrix}

\item{na_zero}{Logical indicating whether NA should be replaced by 0's}
}
\value{
A matrix of 0 and 1
}
\description{
Convert of covariance matrix into a 0/1 correlation map table
}
