% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_common.R
\name{get_derived_parms_code}
\alias{get_derived_parms_code}
\title{Get lines of code for derived parameters}
\usage{
get_derived_parms_code(
  input,
  advan,
  trans,
  isPRED,
  isODE,
  isLINMAT,
  parms,
  parm_lib
)
}
\arguments{
\item{input}{Internal parameter for \code{shiny}}

\item{advan}{Reactive object - NONMEM ADVAN value}

\item{trans}{Reactive object - NONMEM TRANS value}

\item{isPRED}{Reactive object - is model coded with $PRED?}

\item{isODE}{Reactive object - is model coded with ODEs?}

\item{isLINMAT}{Reactive object - is model coded as linear matrix?}

\item{parms}{Parameter selection}

\item{parm_lib}{Library of parameters}
}
\description{
Get lines of code for derived parameters
}
