% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nonmem.R
\name{get_individual_parm_code}
\alias{get_individual_parm_code}
\title{Get line of code for individual parameter value}
\usage{
get_individual_parm_code(parms, varianceTable, iparm, ieta, mu)
}
\arguments{
\item{parms}{Parameter selection}

\item{varianceTable}{Variability selection}

\item{iparm}{Index of parameter in parms data frame}

\item{ieta}{Index of ETA associated with parameter}

\item{mu}{A logical indicator for mu transformation}
}
\description{
Get line of code for individual parameter value
}
