% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_common.R
\name{get_init_code}
\alias{get_init_code}
\title{Get compartment intialization block}
\usage{
get_init_code(input, advan, trans, nPKcmts, nPDcmts, parm_lib)
}
\arguments{
\item{input}{Internal parameter for \code{shiny}}

\item{advan}{Reactive object - NONMEM ADVAN value}

\item{trans}{Reactive object - NONMEM TRANS value}

\item{nPKcmts, nPDcmts}{Number of compartments for PK and PD model components}

\item{parm_lib}{Library of parameters}
}
\description{
Get compartment intialization block
}
