% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mrgsolve.R
\name{get_mrg_individual_parm_code}
\alias{get_mrg_individual_parm_code}
\title{Get line of code for individual parameter value}
\usage{
get_mrg_individual_parm_code(parms, iparm, eparm)
}
\arguments{
\item{parms}{Parameter selection}

\item{iparm}{Index of parameter in parms data frame}

\item{eparm}{Parameter associated with IIV in ordered categorical models}
}
\description{
Get line of code for individual parameter value
}
