% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mrgsolve.R
\name{get_mrgsolve_code}
\alias{get_mrgsolve_code}
\title{Creation of mrgsolve code}
\usage{
get_mrgsolve_code(
  input = NULL,
  template = NULL,
  advan,
  trans,
  isPRED,
  isODE,
  isLINMAT,
  isPREDPP,
  varianceTable,
  covarianceBlock,
  rvTable,
  parm_lib,
  model_lib,
  rv_lib,
  scaling,
  replacement = TRUE
)
}
\arguments{
\item{input}{Internal parameter for {shiny}}

\item{template}{Text template}

\item{advan}{Reactive object - NONMEM ADVAN value}

\item{trans}{Reactive object - NONMEM TRANS value}

\item{isPRED}{Reactive object - is model coded with $PRED?}

\item{isODE}{Reactive object - is model coded with ODEs?}

\item{isLINMAT}{Reactive object - is model coded as linear matrix?}

\item{isPREDPP}{Reactive object - is mode coded with $PK?}

\item{varianceTable}{Variance- table}

\item{covarianceBlock}{Variance-covariance matrix}

\item{rvTable}{Reactive object - residual variability matrix}

\item{parm_lib}{Library of parameters}

\item{model_lib}{Library for $MODEL replacement}

\item{rv_lib}{Library for residual variability replacement}

\item{scaling}{Library for scaling}

\item{replacement}{Logical value indicating with replacement is required}
}
\description{
Creation of mrgsolve code
}
