% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_common.R
\name{get_ncmts}
\alias{get_ncmts}
\title{Get the number of PK and PD compartments}
\usage{
get_ncmts(input, new, model_lib, isPRED, isPREDPP)
}
\arguments{
\item{input}{Internal parameter for {shiny}}

\item{new}{Text template}

\item{model_lib}{Library for $MODEL replacement}

\item{isPRED}{Reactive object - is model coded with $PRED?}

\item{isPREDPP}{Reactive object - is mode coded with $PK?}
}
\description{
Get the number of PK and PD compartments
}
