% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_covariates.R
\name{get_nonmem_blocks}
\alias{get_nonmem_blocks}
\title{Get NONMEM code blocks}
\usage{
get_nonmem_blocks(code)
}
\arguments{
\item{code}{A character string containing the control stream}
}
\description{
Process a character string containing NONMEM model code and returns a
character vector in which each element contain a particular $ block
}
