% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_common.R
\name{get_parms_code}
\alias{get_parms_code}
\title{Get model parameter code lines as list}
\usage{
get_parms_code(input, parms, varianceTable, mu)
}
\arguments{
\item{input}{Internal parameter for {shiny}}

\item{parms}{Parameter selection}

\item{varianceTable}{Variability selection}

\item{mu}{A logical indicator for mu transformation}
}
\description{
Get model parameter code lines as list
}
