% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_common.R
\name{get_preamble_code}
\alias{get_preamble_code}
\title{Get lines of preamble code for transit compartment absorption model and delayed
dosing records}
\usage{
get_preamble_code(input, parms, vars)
}
\arguments{
\item{input}{Internal parameter for \code{shiny}}

\item{parms}{Parameter selection}

\item{vars}{Character vector of variable names}
}
\description{
Get lines of preamble code for transit compartment absorption model and delayed
dosing records
}
