% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_common.R
\name{get_scaling_code}
\alias{get_scaling_code}
\title{Get code lines for scaling and bioavailability}
\usage{
get_scaling_code(input, advan, trans, parm_lib, scaling)
}
\arguments{
\item{input}{Internal parameter for {shiny}}

\item{advan}{Reactive object - NONMEM ADVAN value}

\item{trans}{Reactive object - NONMEM TRANS value}

\item{parm_lib}{Library of parameters}

\item{scaling}{Library for scaling}
}
\description{
Get code lines for scaling and bioavailability
}
