% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_covariates.R
\name{get_theta_number}
\alias{get_theta_number}
\title{Get number of THETA parameters}
\usage{
get_theta_number(code)
}
\arguments{
\item{code}{A character string containing the control stream}
}
\description{
Get number of THETA parameters
}
