% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_covariates.R
\name{hot_to_r_raw}
\alias{hot_to_r_raw}
\title{Convert covariate handsontable data to R data.frame without factor}
\usage{
hot_to_r_raw(...)
}
\arguments{
\item{...}{passed to rhandsontable::hot_to_r}
}
\description{
Convert covariate handsontable data to R data.frame without factor
}
