% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_EDB}
\alias{is_EDB}
\title{Determine if a square matrix is of type block, band or error}
\usage{
is_EDB(x)
}
\arguments{
\item{x}{A square matrix}
}
\value{
Either "error", "band", or "block"
}
\description{
Determine if a square matrix is of type block, band or error
}
