% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_model_server.R
\name{new_model_server}
\alias{new_model_server}
\title{New model module}
\usage{
new_model_server(session, input, output, resources)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}

\item{resources}{A list of internal resources}
}
\description{
New model module
}
