% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{notTruthy}
\alias{notTruthy}
\title{Opposite of areTruthy}
\usage{
notTruthy(...)
}
\arguments{
\item{...}{Any object}
}
\value{
TRUE if at least one object is not "truthy", FALSE if all are.
}
\description{
Opposite of areTruthy
}
