% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxcode-package.R
\docType{package}
\name{pmxcode-package}
\alias{pmxcode}
\alias{pmxcode-package}
\title{pmxcode: Create Pharmacometric Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a user interface to create or modify pharmacometric models for various modeling and simulation software platforms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sbihorel/pmxcode}
  \item \url{https://sbihorel.github.io/pmxcode/}
}

}
\author{
\strong{Maintainer}: Sebastien Bihorel \email{sb.pmlab@gmail.com}

Authors:
\itemize{
  \item Sebastien Bihorel \email{sb.pmlab@gmail.com}
}

}
\keyword{internal}
