% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nonmem.R
\name{replace_abbreviated}
\alias{replace_abbreviated}
\title{Replacement of @ABBREVIATED tag}
\usage{
replace_abbreviated(input, new, vars)
}
\arguments{
\item{input}{Internal parameter for \code{shiny}}

\item{new}{Text template}

\item{vars}{Character vector of variable names}
}
\description{
Replacement of @ABBREVIATED tag
}
