% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nonmem.R
\name{replace_data}
\alias{replace_data}
\title{Replacement of @DATA tag}
\usage{
replace_data(input, new)
}
\arguments{
\item{input}{Internal parameter for {shiny}}

\item{new}{Text template}
}
\description{
Replacement of @DATA tag
}
