% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nonmem.R
\name{replace_des}
\alias{replace_des}
\title{Replace @DES tag}
\usage{
replace_des(input, new, advan, trans, isODE, vars, nPKcmts, nPDcmts, parm_lib)
}
\arguments{
\item{input}{Internal parameter for {shiny}}

\item{new}{Text template}

\item{advan}{Reactive object - NONMEM ADVAN value}

\item{trans}{Reactive object - NONMEM TRANS value}

\item{isODE}{Reactive object - is model coded with ODEs?}

\item{vars}{Reactive object - List of variables in data file}

\item{nPKcmts}{Number of PK compartments in the model}

\item{nPDcmts}{Number of PD compartments in the model}

\item{parm_lib}{Library of parameters}
}
\description{
Replace @DES tag
}
