% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nonmem.R
\name{replace_input}
\alias{replace_input}
\title{Replacement of @INPUT tag}
\usage{
replace_input(input, new, vars)
}
\arguments{
\item{input}{Internal parameter for {shiny}}

\item{new}{Text template}

\item{vars}{Character vector of variable names}
}
\description{
Replacement of @INPUT tag
}
