% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nonmem.R
\name{replace_model}
\alias{replace_model}
\title{Replace @MODEL tag}
\usage{
replace_model(input, new, model_lib, isPRED, isPREDPP)
}
\arguments{
\item{input}{Internal parameter for \code{shiny}}

\item{new}{Text template}

\item{model_lib}{Library for $MODEL replacement}

\item{isPRED}{Reactive object - is model coded with $PRED?}

\item{isPREDPP}{Reactive object - is mode coded with $PK?}
}
\description{
Replace @MODEL tag
}
