% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mrgsolve.R
\name{replace_mrg_capture}
\alias{replace_mrg_capture}
\title{Replacement of @CAPTURE tag}
\usage{
replace_mrg_capture(input, new, parms)
}
\arguments{
\item{input}{Internal parameter for \code{shiny}}

\item{new}{Text template}

\item{parms}{Parameter selection}
}
\description{
Replacement of @CAPTURE tag
}
