% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mrgsolve.R
\name{replace_mrg_cmt}
\alias{replace_mrg_cmt}
\title{Replacement of @CMT tag}
\usage{
replace_mrg_cmt(input, new, model_lib, isPRED)
}
\arguments{
\item{input}{Internal parameter for \code{shiny}}

\item{new}{Text template}

\item{model_lib}{Library for $MODEL replacement}

\item{isPRED}{Reactive object - is model coded with $PRED?}
}
\description{
Replacement of @CMT tag
}
