% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mrgsolve.R
\name{replace_mrg_global}
\alias{replace_mrg_global}
\title{Replacement of @GLOBAL tag}
\usage{
replace_mrg_global(input, new, advan, trans, nPKcmts, parm_lib)
}
\arguments{
\item{input}{Internal parameter for \code{shiny}}

\item{new}{Text template}

\item{advan}{Reactive object - NONMEM ADVAN value}

\item{trans}{Reactive object - NONMEM TRANS value}

\item{nPKcmts}{Number of compartments for PK}

\item{parm_lib}{Library of parameters}
}
\description{
Replacement of @GLOBAL tag
}
