% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mrgsolve.R
\name{replace_mrg_main_pred}
\alias{replace_mrg_main_pred}
\title{Replace @MAIN or @PRED tags}
\usage{
replace_mrg_main_pred(
  input,
  new,
  preamble_code,
  parms_code,
  derived_parms_code,
  scaling_code,
  init_code,
  isPRED,
  parms,
  parm_lib,
  rv_lib
)
}
\arguments{
\item{input}{Internal parameter for {shiny}}

\item{new}{Text template}

\item{preamble_code}{Preamble code}

\item{parms_code}{Typical and individual parameter code}

\item{derived_parms_code}{Derived parameter code}

\item{scaling_code}{Dose scaling and bioavailability code}

\item{init_code}{Compartment initialization code}

\item{isPRED}{Reactive object - is model coded with $PRED?}

\item{parms}{Parameter selection}

\item{parm_lib}{Library of parameters}

\item{rv_lib}{Library for residual variability replacement}
}
\description{
Replace @MAIN or @PRED tags
}
