% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mrgsolve.R
\name{replace_mrg_ode}
\alias{replace_mrg_ode}
\title{Replacement of @ODE tag}
\usage{
replace_mrg_ode(input, new, advan, trans, isODE, nPKcmts, nPDcmts, parm_lib)
}
\arguments{
\item{input}{Internal parameter for {shiny}}

\item{new}{Text template}

\item{advan}{Reactive object - NONMEM ADVAN value}

\item{trans}{Reactive object - NONMEM TRANS value}

\item{isODE}{Reactive object - is model coded with ODEs?}

\item{nPKcmts}{Number of PK compartments in the model}

\item{nPDcmts}{Number of PD compartments in the model}

\item{parm_lib}{Library of parameters}
}
\description{
Replacement of @ODE tag
}
