% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mrgsolve.R
\name{replace_mrg_param}
\alias{replace_mrg_param}
\title{Replacement of @PARAM tag}
\usage{
replace_mrg_param(input, new, parms, nmextImport, posthoc)
}
\arguments{
\item{input}{Internal parameter for \code{shiny}}

\item{new}{Text template}

\item{parms}{Parameter selection}

\item{nmextImport}{A logical indicating whether NONMEM ext file content
should be imported}

\item{posthoc}{A logical indicating whether posthoc estimates should be used}
}
\description{
Replacement of @PARAM tag
}
