% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mrgsolve.R
\name{replace_mrg_plugin}
\alias{replace_mrg_plugin}
\title{Replacement of @PLUGIN tag}
\usage{
replace_mrg_plugin(input, new)
}
\arguments{
\item{input}{Internal parameter for {shiny}}

\item{new}{Text template}
}
\description{
Replacement of @PLUGIN tag
}
