% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mrgsolve.R
\name{replace_mrg_sigma}
\alias{replace_mrg_sigma}
\title{Replacement of @SIGMA tag}
\usage{
replace_mrg_sigma(input, new, rvTable, nmextImport)
}
\arguments{
\item{input}{Internal parameter for {shiny}}

\item{new}{Text template}

\item{rvTable}{Residual variability selection}

\item{nmextImport}{A logical indicating whether NONMEM ext file content
should be imported}
}
\description{
Replacement of @SIGMA tag
}
