% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mrgsolve.R
\name{replace_mrg_table}
\alias{replace_mrg_table}
\title{Replacement of @TABLE tag}
\usage{
replace_mrg_table(
  input,
  new,
  advan,
  trans,
  isPRED,
  nPKcmts,
  nPDcmts,
  parm_lib,
  rv_lib
)
}
\arguments{
\item{input}{Internal parameter for {shiny}}

\item{new}{Text template}

\item{advan}{Reactive object - NONMEM ADVAN value}

\item{trans}{Reactive object - NONMEM TRANS value}

\item{isPRED}{Reactive object - is model coded with $PRED?}

\item{nPKcmts}{Number of PK compartments in the model}

\item{nPDcmts}{Number of PD compartments in the model}

\item{parm_lib}{Library of parameters}

\item{rv_lib}{Library for residual variability replacement}
}
\description{
Replacement of @TABLE tag
}
