% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mrgsolve.R
\name{replace_nmext}
\alias{replace_nmext}
\title{Replacement of @NMEXT tag}
\usage{
replace_nmext(input, new, nmextImport = FALSE)
}
\arguments{
\item{input}{Internal parameter for \code{shiny}}

\item{new}{Text template}

\item{nmextImport}{A logical indicating whether NONMEM ext file content
should be imported}
}
\description{
Replacement of @NMEXT tag
}
