% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nonmem.R
\name{replace_omega}
\alias{replace_omega}
\title{Replacement of @OMEGA tag}
\usage{
replace_omega(new, parms, varianceTable, blocks)
}
\arguments{
\item{new}{Text template}

\item{parms}{Parameter selection}

\item{varianceTable}{Variability selection}

\item{blocks}{Variance - covariance matrix}
}
\description{
Replacement of @OMEGA tag
}
