% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_common.R
\name{replace_path}
\alias{replace_path}
\title{Replacement of @PATH tag}
\usage{
replace_path(input, new)
}
\arguments{
\item{input}{Internal parameter for {shiny}.}

\item{new}{Text template}
}
\description{
Replacement of @PATH tag
}
