% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nonmem.R
\name{replace_pk_pred}
\alias{replace_pk_pred}
\title{Replace @PK and @PRED tags}
\usage{
replace_pk_pred(
  input,
  new,
  preamble_code,
  parms_code,
  derived_parms_code,
  scaling_code,
  init_code,
  isPRED,
  parms,
  varianceTable,
  parm_lib,
  rv_lib,
  mu
)
}
\arguments{
\item{input}{Internal parameter for \code{shiny}}

\item{new}{Text template}

\item{preamble_code}{Preamble code}

\item{parms_code}{Typical and individual parameter code}

\item{derived_parms_code}{Derived parameter code}

\item{scaling_code}{Dose scaling and bioavailability code}

\item{init_code}{Compartment initialization code}

\item{isPRED}{Reactive object - is model coded with $PRED?}

\item{parms}{Parameter selection}

\item{varianceTable}{Variability selection}

\item{parm_lib}{Library of parameters}

\item{rv_lib}{Library for residual variability replacement}

\item{mu}{A logical indicator for mu transformation}
}
\description{
Replace @PK and @PRED tags
}
