% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nonmem.R
\name{replace_prior}
\alias{replace_prior}
\title{Replacement of @PRIOR tag}
\usage{
replace_prior(input, new, parms, varianceTable, estimations)
}
\arguments{
\item{input}{Internal parameter for {shiny}}

\item{new}{Text template}

\item{parms}{Parameter selection}

\item{varianceTable}{Variability selection}

\item{estimations}{Table of estimation tasks}
}
\description{
Replacement of @PRIOR tag
}
