% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nonmem.R
\name{replace_problem}
\alias{replace_problem}
\title{Replacement of @PROB1 and @PROB2 tags}
\usage{
replace_problem(input, new)
}
\arguments{
\item{input}{Internal parameter for {shiny}}

\item{new}{Text template}
}
\description{
Replacement of @PROB1 and @PROB2 tags
}
