% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_common.R
\name{replace_purpose}
\alias{replace_purpose}
\title{Replacement of @PURPOSE tag}
\usage{
replace_purpose(input, new, variance)
}
\arguments{
\item{input}{Internal parameter for {shiny}.}

\item{new}{Text template}
}
\description{
Replacement of @PURPOSE tag
}
