% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nonmem.R
\name{replace_sigma}
\alias{replace_sigma}
\title{Replacement of @SIGMA tag}
\usage{
replace_sigma(new, input, rvTable)
}
\arguments{
\item{new}{Text template}

\item{input}{Internal parameter for {shiny}}

\item{rvTable}{Residual variability selection}
}
\description{
Replacement of @SIGMA tag
}
