% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nonmem.R
\name{replace_subroutine}
\alias{replace_subroutine}
\title{Replacement of @SUBROUTINE tag}
\usage{
replace_subroutine(input, new, advan, trans, isPRED, isODE, isLINMAT)
}
\arguments{
\item{input}{Internal parameter for \code{shiny}}

\item{new}{Text template}

\item{advan}{Reactive object - NONMEM ADVAN value}

\item{trans}{Reactive object - NONMEM TRANS value}

\item{isPRED}{Reactive object - is model coded with $PRED?}

\item{isODE}{Reactive object - is model coded with ODEs?}

\item{isLINMAT}{Reactive object - is model coded as linear matrix?}
}
\description{
Replacement of @SUBROUTINE tag
}
