% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nonmem.R
\name{replace_table}
\alias{replace_table}
\title{Replace @TABLE tag}
\usage{
replace_table(input, new, vars)
}
\arguments{
\item{input}{Internal parameter for \code{shiny}}

\item{new}{Text template}

\item{vars}{Reactive object - List of variables in data file}
}
\description{
Replace @TABLE tag
}
