% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nonmem.R
\name{replace_tags}
\alias{replace_tags}
\title{Replace $TAGs tag}
\usage{
replace_tags(input, new)
}
\arguments{
\item{input}{Internal parameter for {shiny}}

\item{new}{Text template}
}
\description{
Replace $TAGs tag
}
