% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nonmem.R
\name{replace_task}
\alias{replace_task}
\title{Replace @TASK tag}
\usage{
replace_task(input, new, estimations, isODE)
}
\arguments{
\item{input}{Internal parameter for {shiny}}

\item{new}{Text template}

\item{estimations}{Table of estimation tasks}

\item{isODE}{Reactive object - is model coded with ODEs?}
}
\description{
Replace @TASK tag
}
