% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_common.R
\name{scale_value}
\alias{scale_value}
\title{Prepare initial estimate, min, and max for $THETA}
\usage{
scale_value(min_value, value, max_value, scale)
}
\arguments{
\item{min_value}{the minimum value entered in the UI}

\item{value}{the value entered in the UI}

\item{max_value}{the maximum value entered in the UI}

\item{scale}{the scale entered in the UI}
}
\description{
Prepare initial estimate, min, and max for $THETA
}
