% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mrgsolve.R
\name{get_mrg_parms_code}
\alias{get_mrg_parms_code}
\title{Get mrgsolve model parameter code lines as list}
\usage{
get_mrg_parms_code(input, parms, mu)
}
\arguments{
\item{input}{Internal parameter for {shiny}}

\item{parms}{Parameter selection}

\item{mu}{A logical indicator for mu transformation}
}
\description{
Get mrgsolve model parameter code lines as list
}
