% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxcode-package.R
\docType{package}
\name{pmxcode-package}
\alias{pmxcode}
\alias{pmxcode-package}
\title{pmxcode: Create Pharmacometric Models}
\description{
Provides a user interface to create or modify pharmacometric models for various modeling and simulation software platforms.
}
\author{
\strong{Maintainer}: Sebastien Bihorel \email{sb.pmlab@gmail.com}

}
\keyword{internal}
