% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nonmem.R
\name{replace_theta}
\alias{replace_theta}
\title{Replacement of @THETA tag}
\usage{
replace_theta(new, parms)
}
\arguments{
\item{new}{Text template}

\item{vars}{Character vector of variable names}
}
\description{
Replacement of @THETA tag
}
